Implement better hashing and checking of the RSS item guid
Was doing a crap job of checking for uniqueness in the first version of the plugin and I'm not convinced it ever got much better. This does an MD5 hash of the RSS item guid and uses that for uniqueness. Now a title can change and we won't know create a new item.

Of course, now that I type that out, maybe we want to detect title changes and update the post information if something has changed. Or maybe not. We'll see.